+++ Test/Hspec/Server.hs
@@ -1,13 +1,15 @@
 {-# LANGUAGE OverloadedStrings #-}
 module Test.Hspec.Server (
-  module Test.Hspec.Server.Type
-, module Test.Hspec.Server.Core
+  module Test.Hspec.Server.Core
 , module Test.Hspec.Server.ServerType
-, module Test.Hspec.Server.Command
+, module Test.Hspec.Server.ServerStatus
+, module Test.Hspec.Server.CommandStatus
+, module Test.Hspec.Server.NetworkStatus
   )
 where
 
-import Test.Hspec.Server.Type
 import Test.Hspec.Server.Core
 import Test.Hspec.Server.ServerType
-import Test.Hspec.Server.Command
+import Test.Hspec.Server.ServerStatus
+import Test.Hspec.Server.CommandStatus
+import Test.Hspec.Server.NetworkStatus