+++ ChangeLog.md
@@ -1,3 +1,8 @@
+## 0.4.0
+
+* Add NetworkStatus and function for checking network-reachablity
+* Change ServerStatus-Type to Data.Set ServerStatus' and Remove Sets-type-class
+
 ## 0.3.2
 
-Support for docker
+* Support for docker